Added executable attribute to linter class.
This allows the linter to be enabled/disabled based on the availability of the linter's executable, without overriding a method.

If set, it should be the plain name of the executable. If the executable is in the user's path, executable_path is set to the full path. If it can't be found, the linter is disabled.