Fixed: eliminate race condition between main thread and linting thread.
Previously, when a lint was finished, it's marks were always applied to the view. But if the view was modified during the asynchronous lint, the marks could be invalid.

Now a timestamp is recorded with the lint, and if the view has been modified since that time, marks are not applied.

Also, marks are always cleared now whenever the view is modified.