Fixed Highlight.near() to be smarter
In a line like this,

   if foo['chooser'] == chooser:

if chooser was an undefined name, the string 'chooser' was selected by Highlight.near().

Now, near() checks to see if the text to match is enclosed in quotes. If so, a simple find is done. Otherwise a regex is used to match the first occurrence of the text that is not enclosed in quotes.