Fixed: user settings weren't available at startup.
Previously queue.start was called in SublimeLinter.__init__, which called util.get_python_paths, which called util.create_environment, which tried to get the user settings but failed because they aren't available until after all plugins are loaded.

Now the queue is not started until plugin_loaded to ensure the settings are available.