Adds make_warnings_subtle setting
Alternative fix #89

Setting `"make_warnings_subtle": true` in your
SublimeLinter.sublime-settings will now cause warnings to not
display when using the Show All Errors command.

This is the same command used when `"show_errors_on_save"` is set
to `true`, so enabling it will also cause warnings to not show
in the popup errors panel on save.