Adds hide_warnings_from_all_errors setting
Setting `"hide_warnings_from_all_errors": true` in your
SublimeLinter.sublime-settings will now cause warnings to not
display when using the Show All Errors command.

This is the same command used when `"show_errors_on_save"` is set
to `true`, so enabling it will also cause warnings to not show
in the popup errors panel on save.