Fixed: On Windows, subprocess.PIPE fails without stdin.
As of python 3.4.0 and earlier, when calling subprocess.Popen on Windows with stdout or stderr == subprocess.PIPE, nothing is returned if nothing is sent to the process via stdin. This would cause linters that do not use stdin to fail, e.g. jscs.

This commit works around the bug by sending stdout/stderr to a temp file and then reading the output from the temp file.

Closes https://github.com/SublimeLinter/SublimeLinter-jscs/issues/3
Closes https://github.com/SublimeLinter/SublimeLinter-jscs/issues/6