Fixed incorrect description of what keys in syntax_map correspond do
Making this change after I tried to match a custom syntax definition by what was shown in the corner of the status bar, and it would not work at all. It turns out that the .tmLanguage file that defines the syntax had a different name (e.g. the status bar was showing "Custom Syntax", but the file was named "Syntax Custom"). After changing syntax_map to have a key of "syntax custom", it worked perfectly.