Override can_lint for NodeLinter
Without this change, if the linter executable is not found globally (with `util.which(cmd)`), the NodeLinter will be disabled.  Since `can_lint` is called with the view syntax only (and not the view), and the executable used will vary based on the view filename, the NodeLinter's implementation is optimistic, determining its ability to lint based on syntax alone.