add config_joiner setting in linter class
This setting can be set to either  '=' or ':'.

The default setting of  ':' makes the config file argument and the value passed as separate arguments for backwards compatibility.

When this setting is set to '=', the config file argument is joined with the config file path found by '=' and passed as a single argument.