Disable Panel Cancel Branch in select_error
If you don't actually click on something in the panel then the quick_panel cancellation branch of the if statement causes the window to jump around and messes up where you are clicking. Disabling this will probably change how things work if you click on an item in the panel, but I never do that.