Add `use_rvm_auto_ruby` setting
By default, SublimeLinter will try to use `rvm-auto-ruby` executable
for linting. Although this guarantees that the latest RVM-activated ruby
will be used, it has the downside of being too slow (it re-initializes
the RVM environment on every invocation, which has a noticeable
performance impact).

RVM installed should not mean that `rvm-auto-ruby` is the only means
of executing RVM-enabled ruby. In my case, I do have RVM, but I
initialize it only once, at Sublime boot, so paths are already in place
and the linter will lint using the "correct" ruby. See this gist:

http://gist.github.com/smanolloff/54a787f661a3b8b5af10541f5cc5fa2c