Remove "message" group in line_regex again
The matched error message shows up in the status bar when using F4 to
navigate the errors, but was displayed by SL from the caret being in the
same line as well.
We don't need both, so remove the match again.

A more specific regex was included as a comment to strip the white
spaces that were kind of obscuring the actual message.