Diagnostics panel improvements
This commit enables the gutter area of the output panel and indents each line by a single `tab` in order to enable ST´s code folding capabilities. This way a file which is out of interest can be collapsed. In order to ensure correct indention, the line:col pair is moved to the second column and the severity to the first one and displayed as unicode character. As a result the empty line between each file seems no longer required and is therefore removed. To ensure code folding works properly the very last line must end with a `\n`.