Remove queue.Empty exception from daemon traceback
All errors occurring in linters were encountered within the event
handler of the `queue.Empty` exception, which caused the printed
traceback to be unnecessarily long and confusing, since the part is
entirely redundant.
Since the exception here is used in a specific way (almost in a way that
it is the expected event), I instead moved the code from within there
outside of the exception handler and the code from outside into an else
block.