Queue replacement
The old queue implementation used its own `threading.Thread` to do async
background work. It had to use polling for that.

Here, we use the builtin async threading that come with Sublime. This has
the advantage that we respect the event system of Sublime. Sublime may
decide that it is busy right now, and will therefore delay our workers.

It is also an easier and shorter implementation.

A small trick is applied for startup so that the initial linting of the
visible views actually occurs even without user interaction.