Fix: Use PROTECTED_REGIONS_KEY once
Region keys must be unique per view, so a set is the natural type here.
By doing that I found that we actually use PROTECTED_REGIONS_KEY twice
per draw.

So, instead of using a set, I kept the list and check if we actually
hold the promise to keep the keys unique.