Re-enable `self.tab_width` setting, document `reposition_match`
`reposition_match` was a really important functionality hidden inside
Highlight. It fulfills the purpose that SublimeLinter finds a range or an
end for each match. (Linters usually just provide the starting point via
`col` for an error. )

By putting this function in linter, users could extend this functionality
and 'reposition' the found error.