Remove ${root} token
Since we try to reduce and remove the bugs from ${project}, we hope there
is no need for ${root}.

Basically reverts PR #581

Also the original author wanted to apply this to `chdir`, and we decided
to do this by default, exactly how he proposed. So, the good news would
be, remove your setting, we have it as default in core.