Specify working directory in Popen call
Using os.chdir is a bad choice and thread-unsafe, which bites us
when running linters concurrently.
To ensure thread safety, this function must be removed entirely,
which is a breaking change for linters using it.

I omitted tmpdir here because I have no idea how this is used
(if at all).