Fixes #912 where a file extension is required for some linters
I found that some linters like SublimeLinter-phpcs [1] stopped working
after a recent update. After some debugging of the code I found that
most recent versions of PHP CodeSniffer do not work when the file has a
different extension than `.php` or has no extension at all.

Before SublimeLinter executes the command specified by the variable
`cmd` it creates a temporary file, which on macOS is placed at `$TMPDIR`
and then copies the content of the original file there, then it executes
the command using `process.Popen`; the problem here is that the files
are being created using `tempfile.NamedTemporaryFile` which by default
creates file names with this format `/tmp[a-z_]{6}`.

A solution is to pass the `suffix=".ext"` parameter to the method call.

[1] https://github.com/SublimeLinter/SublimeLinter-phpcs