Utilize logging stdlib under the hood
Use formatters for the individual loggers to achieve different prefixes.
Also respond to updates of the view's file name by re-creating the
logger to assign a new formatter.

`print` calls have been replaced with `info`, `warning` or `error`
respectively. For now, the log level is not printed from the formatter
though to maintain current behavior. This may be changed later.

Note that log records are formatted using the `%` string format operator
and this is not configurable without affecting global state.