Modify Linter.get_tempfile_suffix to extract file extension (#917)
Right now, the `Linter.get_tempfile_suffix` method is returning an empty
string as a fallback mechanism when the extension for the temporary file
is not found. This commit adds an attempt to extract the extension from
the filename (if available) and defaults to an empty string if the file
has no name.