Change signature of `context_sensitive_executable_path`
BREAKING: make fn `(str) -> Optional[List[str]]`

`context_sensitive_executable_path` had a polymorphic signature in that
it also allowed the full `cmd` to be passed in. However all
implementations then only extracted `cmd[0]` and acted upon that.

Esp. for the new `which_executable` we want a simpler signature, so that
the user can just call e.g. `self.which_executable('mypy')`. We, of
course, could transform this internally to `['mypy']`, pass it down,
and transform it back to `'mypy'` in every `csep` implementation.