Focus on active file in the panel
Fixes #1415

We try to wrangle the panel output so that it appears as if the panel
only shows the errors from the current file. For that, we just put the
focused content/file at the end of the panel output and scroll to it.

This has the advantage that we don't loose the other errors. We also don't
fool the user by saying there are only these errors. We just visually
highlight the current file.

For clean files, we show "No lint results." in place of errors.