Start or stop viewport poller on_activated
The poller currently is global. If the user switches from one window
to another one, we can stop the poller if the active window actually
has the panel not open, if she then switches back to a window with an
open panel we start the poller again.

This solves two issues:

- If the user switches to a new window `fill_panel` actually raises
  (if called basically random, at will via the poller) because the
  panel has no size yet.

- If the user has multiple windows, all with the error panel open, the
  poller stops running as soon as the user closes one of them, making
  the panel of the second window disfunctional.