Fix `disable_if_not_dependency`
Fixes #1559

The setting `disable_if_not_dependency` was disfunctional. If a user set it to
`true` the status bar still indicated '(erred)'.

Now, there aren't any good solutions here so we dig a new hole. Generally we
raise during linter execution to rewind the stack. We introduce a new error
`PermanentError` as a compagnon to `TransientError`, and define new side-effects
for the status bar.

- If we receive a `TransientError` we keep the old errors, and the old status
  bar information. To achieve this we re-raise the/an error in the backend
  which transform into a `None` which in turn signals to *NOT* call the
  continuation `sink()`.

- If we receive a `PermanentError` we erase old errors, and remove the linter
  ('unassign') from the status bar. If we don't do the latter we would get an
  '(ok)'. To achieve this we return `[]` in the backend and call the `sink()`
  normally. This will update the error repo and with it e.g. the panel. To not
  get a '(ok)' we introduce a new event `sublime_linter_unassigned` which we
  send to the status bar.

We clearly see that 'active_linters_view' and the backend almost wants to be
coupled, and we don't want that so at least we only send events. Note for the
future: The messages still always flow from backend to view.

Ideally we could just use `can_lint_view` for that feature but it - as a class
method - doesn't have access to the information we need to make the decision.