Ensure independent, mutable linter instances in backend
Fixes #1564

People are used to the OOP rule that if they see a `self` they can mutate
at their will. This works okay if the linter selects the whole buffer because
then the linter is used only once. If the linter selects multiple regions
instead we always reused the linter which can lead to subtle bugs.

We therefore clone the linter if the linter selects multiple regions.

We ensure that cloned linters inherit all modifications made to the settings
from the 'parent'. E.g. a linter could modify the settings in `can_lint_view`,
and these changes must be visible to the cloned linters.

All changes after the cloning are local.