Make `LintMatch.match` optional
Fixes #1562

We do not use `LintMatch.match` internally, and for JSON reporters it is
annoying to pass a truthy value here just to get the error passed.

We implement here easy as **breaking change** because it is unlikely users
explicitly set the first item of our 7-tuple to `None` but have valid values
for the other holes.