Improve NodeLinter lookup algorithm
Generally the happy path is to find just a binary in './node_modules/bin'.
If that fails we look for a 'package.json' file, and read it.

We now support the edge case that the user might hack on the linter itself.
(We "had" this before but the implementation didn't work.)

We skip the linter if we find a dependencay on it which is not installed.
(Prior it grabbed the global eslint.)

Note: for all failures here, we call `notify_failure` and additionally raise
`PermanentError`. This has the effect of cleaning all errors and putting an
'(erred)' in the status bar.