Efficient use of `ThreadPoolExecutor`
The prior implementation did not reuse the `executor` which is a waste. To
actually limit the number of threads, we additionally had to use a
`BoundedSemaphore`.

We replace this system with two global executors. We need *two* executors
because we still have a nested structure here. The orchestrator does the
fan-out, fan-in dance for each linter in case the linter has multiple regions
to lint. The executor just runs the long lint tasks.