Refactor opening the panel after saving
Just asking for the global 'lint_mode' is not enough since all linters can have
their own settings here. See: #1581

It looks much simpler if we just pass down the reason when we propagate the
lint result because we get rid of the 'just_saved_buffers' state as well.

Generally, event listeners should be open for addition (using `**kwargs`) so
this should be save. Note that we invent a new reason 'relint_views' on
configuration changes, this works because `should_lint` is also open for new
reasons.