Disable linter with incompatible overrides
Disable linter which implement `get_environment` or `get_working_dir` using the
old signature as there is no compatibility layer for that and they would
otherwise just throw badly.

Note: We only ship a compatibility layer for calling or using these methods
which is by far more likely. We actually did not found a public instance
of a plugin *overriding* said methods.