Just silently (and locally) assign a linter if we get a report
If we get a lint result with actual errors we can assume the linter
is applicable for that filename. This is only the case for multi-file linters
reporting filenames with either no views or never acivated views.

Generally: 'assigned' is fired on first activate, and re-checked after every
modification for all views. ('assigned' gets called even if no linters will
actually run; it's meaning is: we have linters that could *possibly* run for
this view.)

On the other side: 'unassigned' and 'failed' are events that only fire if
a running linter fails loud (or rather silent in case of 'unassigned').
(T.i. these are RuntimeErrors)