Remove clear errors command
The command was useful for testing, we could also revert this removal later.
But for now, we want the cleanup to be as automatic as possible. We actually
don't want people to invoke any commands at all. By removing it, it is much
more likely people report annoying issues where errors remain in the panel
when they should disappear.