Redo executable lookup in `NodeLinter`
We now traverse up (usually coming from the dir of the current filename)
looking for 'package.json' files. Whenever, we find one, we look into it
and see if the wanted `cmd` is declared as a dependency. If not we continue.
If yes, we take that dir and set it as `context['project_root']` which we
later readout as the `working_dir`.

For the upcoming yarn pnp, we look for markers (e.g. a '.pnp.js' file) and
in case just run `yarn run -s cmd`. (If `yarn run` fails, we show an error
message after the fact, basically being optimistic here.)

Fixes #1626
Fixes #1644