Implement cache on `settings` object
We completely cache the settings in memory. This surprisingly boosts
performance because if we access `load_settings` from the worker
or any other thread we often pay a synchronization penalty of
multiple ms. (Then again these lookups are zero cost.)

The old implementation was also too sloppy because `has_changed`
could only be called once in `on_update` per key. (On the second
call it returned always `False`.)