Show affected files below the active one
Fixes #1582

This is really just mangling with sorting.

With one exception: If the current file is clean, it doesn't mean
anymore it is also the last entry in the panel. (Because after the
current file we will show its dependencies.) For now, we just kind
of brute force search for the literal "No lint results" message
and scroll to the line before that because that's where the filename
is.