Maybe resolve executable set by user
Fixes #1635

We actually only ever expected absolute paths in the 'executable'
setting. Of course, users try short executable names here, and
wonder if it fails.

'executable' can be a plain string or a list of strings. In the latter
case, the real executable is the first item in the list.

We now either check if the given 'real' executable is there and
actually executable if it is an absolute path, or try to resolve
it with the usual `which` call.

As always, the error handling and messages make everything relatively
long.