Ensure file properties in the context come from given view
Fixes #1690

We use `window.extract_variables()` and took it as-is for the context.
We knew we had to overwrite all file properties to get some consistency,
but we never *erased* possible file properties if the given view is
actually unsaved.

That's a bug which can be exposed *iff* the linter uses e.g. `$file`
as part of its command. The impact is relatively small because at this
time basically no linters report filenames still, and then the backend
just fills in the correct canonical filename by itself. Otherwise,
"saved-file-only" linters would have shown the bug early but they
obviously never run on unsaved, dirty views.

Typical info in the console looks like

```
> type <buffer 8869> | eslint.CMD --filename "...\linter.py"
```

indicating that buffer `8869` is linted with the name `linter.py` which
is inconsistent.

The fix is to just remove all file related properties from
`extract_variables()` unconditionally.

Note:

Avoiding the race between getting the view (in the event handler) and
actually freezing its context is impractical:

1. We don't want to do this on the main thread because it would be a
   waste.
2. We don't want to abort if we detect this inconsistency, we really
   just want the fixed context in that case which leads to the same
   code as here.