Ensure `active_view` is from given window
Fixes #1693

We currently only store the `active_view` and `active_filename`
globally; but we can have a panel per window.

This change ensures the `active_filename` comes from the window
we want to draw otherwise we possible show "No lint results." for
a file actually belonging to a different window.