Log output of a failed command
in addittion to the exception output of subprocess.check_output.

Ref: https://docs.python.org/3.5/library/subprocess.html#subprocess.CalledProcessError.output

I found it very useful to see the full output in debug, as was for me
when checking an error that occurred due to wrong version of python.
Pipenv (or rather it's dependency "click") when run on macos 10.15 is
missing some required "locale" env variables in sublime, which was
fixed in versions 3.7.5 / 3.8. Ref: https://bugs.python.org/issue18378