Mark overlapping errors better in the panel
For errors sitting on top of each other (but i.e. with different
lengths) we now highlight/mark all of them. Being an edge-case
of course, consider this:

     here.comes_an_error
     ^^^^                  linter A
     ^^^^^^^^^^^^^^^^^^^   linter B

           |               cursor C
      |                    cursor D

Behavior before this change:

Cursor at C:

Before this change we highlight (surprisingly) the error from linter
A! Why? Since both start at the same position, the distance from the
cursor to the start is the same for both, we then take the end into
account, and here the distance to the end of A is smaller. Since region
A does not *contain* B, we only select A.

Cursor at D:

Since the distance from D to the beginning of A or B is the same, we
choose one of them sort of randomly. If we choose A, we again do not
select B as well because the region of A does not *contain* B.

After this change:

Cursor at C:

We just select B because it gets an extra point since the cursor is
within its region. We do not select A, because the region of A does
not contain B.

Cursor at D:

We select A because it's shorter. When we select A, we also select
B because B contains A.