Really remove all errors associated with empty regions
Although a real Dangle! is empty and at col 0, users can produce empty
regions for example by just deleting the squiggled code.

Before we kept these errors in the store and with it e.g. in the panel.
On the other side there are useless because you cannot really jump to
them anymore. We will now remove them when we catch'em. The only down-
side probably is that then the panel is a bit more dynamic just in the
sense that it updates immediately on user press instead of after a lint
process ends.

(Rarely these user created dangles stayed and became even Zombies so
removing them immediately we don't have to find that race condition.)