Continue spinning if the user activates a special view
Fixes #1727

The busy spinner immediately stops spinning if the user opens or
activates a different view. This is only the right thing if the selected
view is a non-special view.

The fix is to check `util.is_lintable` before accepting the new view
as the "active" one. `is_lintable` filters all panels for us.

We switch to `on_activated` (t.i. the version on the UI thread) because
this reduces status bar repaints (flicker) if you really `<tab>` to a
new view while the spinner is showing.

(Basically being "sync" here we avoid one repaint; without it Sublime
would first update the status bar info to match the new view, then we
would kick in and update the spinner; with it we do this in one paint.)