Allow going to the next file
If a file has "dependent" files with errors we show them right below
the current file.  In that case (and only then) it seems very reasonable
to move to these next errors, probably opening transient views.

This is relatively straight forward to implement.  Since we already had
implicitly the notion of a "top" file, which was always the active file,
we now set the top file to the file the navigation started with, t.i. the
"original view".  (Simply put, "follow active view" is off in quick panel
mode.)  We do this because otherwise, upon navigation, the panel would
wildly redraw, always showing the active file on top.
Note that we don't allow navigating to errors *above* the orignal, top view.

Noteable for "commit":
We assume that the cursor is already in the correct place.  We actually
trigger `open_location` to maybe turn the transient into a normal view,
and in any case put the new location in Sublime's navigation history.
But we don't actually jump anymore or read the location a user might have
*clicked*.
T.i. we remove the click + `<enter>` feature for now.

Partially reverts c673a6da00ff19480f21de9dbf4b3d309c9ec705 ("Implement
keyboard navigation for the error panel")

This is now unused since we switched to `open_location` solely.