Simplify `scroll_into_view`
Since a not so recent change we actually show the "active" file always
on the bottom of the panel.  It is thus not necessary to compute a
`fbottom` (file bottom = last line of the file) as a bottom boundary
the scroller must not overscroll.  Instead the bottom is always the
last drawn line of the panel.  (Still computing with the `fbottom` is
nice because Sublime can overscroll and show more whitespace at the end
of each view. We avoid that here and stick the last line to the bottom
of the viewport.)

With that we can simplify the types here as well.  We don't send all
`errors_from_active_view` around but instead just the one line of the
filename.  Unfortunately, if there are no errors for a particular file
we still need to query the panel/view for a good line.