Do not use `Default.history_list` for ST4
Fixes #1763

In ST4 the behavior of the Jump History changed and with it
`Default.history_list.get_jump_history_for_view` has been removed.

We apply a quick fix here because of the immediate breakage.

Note that "add_jump_record" only adds a JumpRecord if more than 1 sec
has passed since the last call.  So our feature is broken in ST4.