Allow for customization of Quick Fixes
This commit implements the capacity to configure the Quick Fixes
made available by SublimeLinter. This commit deals solely with the
ESLint Quick Fixes but lays the foundation for extending the
customization to the other linters that can apply the Quick Fix feature.

Taking ESLint as our example, it is personal preference whether an
ESLint violation is ignored with an `// eslint-disable-next-line
VIOLATION(S)` or an `// eslint-disable-line VIOLATION(S)` comment.
Currently, SublimeLinter only provides the functionality to use the
`// eslint-disable-next-line` comment. It would be preferable to be able
to configure this to use the type of inline Quick Fix that the
programmer prefers.

TODO - Actually detail the final SublimeLinter settings