Remove `end` key from `LintError`
Specifying a "start" and "end" value for an error has been superseeded
by the use of "region".  We only use "line"/"start" as a cached value
for display in the panel.

Other than that, "end" is used for sorting as there can be more than one
error at a given start position.  `len(region)` or `region.end()` work
here as well.

Notice that we remove a lot of finalizing in `backend.finalize_errors`
which is okay as the code is unused in the wild, and not covered by any
tests.  Of course, we ideally have only *one* place where we "finalize"
errors and this is now `process_match`.  In a future refactoring, it is
advisable to move `process_match` to the `backend`.