Simplify normalizing `line`, `col`, etc.
This has a change in behavior though it is not covered by the tests.
Before if `match.groupdict()` for example had no key `col` the resulting
`LintMatch` followed.  Now we add that key setting the value to `None`.

Typically, the match objects are used with dotted notation, and
`match.col` returns `None` in both cases, the key not being present *or*
the value being `None`.

As user you don't want this doubled optionality anyway.