Add `self.logger` to the linter classes
For convenience, add a `self.logger` to all linter classes so that users
don't have to `import logging; logger = ...` at their plugins.

Especially setting the correct namespace for the logger was difficult
for end-users, and I always had to lead them in the right direction.

Note that we also use this namespaced logger within our code finally
as the "logging" approach has been quiet succeful.