Remove usage of `verifyNoUnwantedInteractions`
As mockito has been updated, it checks automatically on `__exit__`.

Two usages were misleading, as the high level `execute_lint_task` eats
all exceptions, it is always necessary to check afterwards to fail the
test. In this case the classic `when/verify` is probably what you want,
as a `with expect...` doesn't work as expected.