Fix: Draw dangle and icon in EOF position
Fixes #1784

We always ensured regions of size 1 for all errors, because it looks
better to not have dangles all over the code.  We further recognize
empty regions which can happen when the user edits on errors, or deletes
or moves blocks of line, and remove these automatically.  That means
an empty region has a special meaning in SublimeLinter.

Unfortunately, Sublime will not draw squiggles for regions that go past
the EOF at all as #1784 has shown.

Let the linter produce empty regions for that exact edge case on EOF,
t.i. exactly the region `Region(view.size())`.  However, if only parts
of a file are linted, the virtual view inside the linter can "think" it
has an error on eof when it is actually on the end of the linted *part*
of the file only.  Check in the backend, and maybe undo what the linter
did.

Draw empty regions as dangles, with the flag `DRAW_EMPTY_AS_OVERWRITE`,
and ensure they're excluded from our dangles/ zombie protection.