Add generic enabled switch
Fixes #1848
Fixes #545

Introduce a special key `SublimeLinter.enabled?` read from the view
settings which is a 3-state toggle (`null, true, false`).

This can be set to `True` for specific views as requested by #1848 to
"force" lintability of them.

On the other hand it could be set in the project settings, for example
to `False`, to toggle SublimeLinter off completely for that project.