Fix: Allow setting keys with funny suffixes
Setting keys can have a special format and then SublimeLinter will
treat them as arguments to the linter binary.

The logic in `build_args` for this is that the setting must have a
prefix of either `@`, `-`, or `--`.  Unfortunately, we did not follow
this logic when building the linter class in `LinterMeta`, but instead
we threw away some settings.

E.g. `enabled?` does match our pretest regex because of the trailing
`?` but fails the later test because it has no prefix.

Fix this by testing for a prefix in `LinterMeta` as well.