Put error "code" at the end of the first message line
Place the error code right-aligned at the end of the first line from
an error.  Use an invisible space to (instead of the `:`) to mark the
beginning of this label to make matching the syntax trivial.

This leads generally to more space for the message as subsequent lines
can fill up to the width of the viewport.  It is also especially
preferable when the linter has very long rule names.