Handle `BrokenPipeError`'s on Windows
Backport `Popen._communicate` from new python versions to py3.3.  In
Python 3.3 we have a `self.stdin.close()` call which is completely
unprotected and throws if the process already finished.  We shouldn't
care about `stdin` but recover and read the info possibly present on
`stdin` and `stderr`.

This has been discovered by using the ultra-fast ruff linter.