Do not consider `HOME` as a project root
We used `smart_paths_upwards` which used to `yield HOME`.  But that was
a special case for `NodeLinter` which doesn't look for markers but for
executables.

Change the function design so that `paths_upwards_until_home` goes
upwards halting either before HOME or, on non-Windows systems, the root.

This is the right thing for `PythonLinter` as we look for marker files,
typical configuration files, to decide which directory could be the
project root.  And it is not uncommon to have a "global", fallback
`.flake8` configuration in `HOME`.