Redo logic how we expand or shrink the status bar information
Fixes #1809

The infinite loop from #1809 comes from calling a state modifier in
`draw`.  That's just confusing.  Instead, enqueue the "unsetter" always
after the "setter", because that is what we want:  we want `expanded_ok`
to be a temporarily set for new views and whenever the assigned linters
change.

`expanded_ok` just means:  *force* a verbose status message even if the
view is clean.