Fix minor issues with the flake8 quick fixes
In falke8 we change the position of E302, E303, and E305.  For the
quick fix, we need to undo these changes otherwise we ignore on the
wrong line.

(Maybe this should be moved to the flake8 plugin, that would be much
clearer, and ruff and eslint already have their own fixers AFAIK.)