Set `VIRTUAL_ENV` vars
If a virtual environment is found, set `VIRTUAL_ENV` and augment `PATH`
to mimic exactly what the activation scripts do.

Un-deprecate `self.env` for this information as it is merged
low-priority in `get_environment()`.  (T.i. a user can override our
finding.)

Note that the comment on `self.env` about "surprising side-effects for
concurrent(...) linting" was already wrong as we don't share linter
instances anymore.  We always and only use them once.