Humanize column in the panel
We literally used `rowcol(error.region.a)` to compute the line and start
used in the panel.  This is ok, but if the user uses real tabs, the
column we show does not match the visual column the user sees in the
editor.

Actually, Sublime Text shows the humanized/visual column in the status
bar, and that mismatch makes it even stranger.  Once noticed you cannot
unsee it.

Fix it by reading the "tab_size" and adjusting the `rowcol`.

Note that "tab_size" belongs to the "context" of the lint request, just
like all the other properties of the view.

Rel SublimeLinter/SublimeLinter-phpcs#57