Revert "Humanize column in the panel"
From the perspective of Sublime Linter the "start" (or: "column") value
is just a presentation value, precomputed for display in the panel.

However, the literal text in the panel is a "data" value when we bind
"result_line_regex" to drive the global "F4" (`next_result`) location
list.  For this feature, we can't have "humanized" columns as Sublime
Text does not support this here.

We can at a later point fix this if we want to but do note that the
behavior of the status bar column display is actually determined by
a setting

```
// Controls what type of columns selection description ("Line X, Column Y"
// in the status bar) uses:
// - "virtual": Count virtual columns. Tabs count as their width in spaces.
// - "real": Count real columns. Tabs count as one column.
"selection_description_column_type": "virtual",
```

which we did not respect in the first version.

This reverts commit 22a0b52a9d3250b441ca6ad7a845dd0398fac738.